docker/backend/requirements.txt
flask==2.3.2
redis
psycopg2-binary
prometheus-client
gunicorn
